Вынести обработчик ошибок и rate-limiter в отдельные файлы
