+++ controllers/articles.js
@@ -28,6 +28,7 @@ const createArticle = (req, res, next) => {
       if (err.name === 'ValidationError') {
         throw new BadRequestError('Переданы некорректные данные');
       }
+      throw err;
     })
     .catch(next);
 };
