+++ middlewares/auth.js
@@ -6,7 +6,7 @@ const { SECRET_KEY = 'eminem' } = process.env;
 module.exports = (req, res, next) => {
   const { authorization } = req.headers;
 
-  if (!authorization && !authorization.startsWith('Bearer ')) {
+  if (!authorization || !authorization.startsWith('Bearer ')) {
     throw new UnauthorizedError('Необходима авторизация');
   }
