+++ middlewares/rate-limiter.js
@@ -1,0 +1,9 @@
+const rateLimit = require('express-rate-limit');
+
+const limiter = rateLimit({
+  windowMs: 15 * 60 * 1000,
+  max: 100,
+  message: 'Превышено допустимое количество запросов!',
+});
+
+module.exports = limiter;
